feat(43-06): convert /hub to RSC + wire viewMode through HubResults
- app/hub/page.tsx is now a server component reading hub_view cookie
  via cookies() from next/headers; default 'cards' on miss/invalid
- app/hub/_view-shell.tsx (NEW, client) holds the existing hub page
  body with HubViewToggle inserted in the header next to Sort
- HubResults accepts required viewMode prop; list mode falls through
  to the cards grid with TODO(43-07) for the real WorkflowTemplateRow
- data-view attribute on root + data-view-mode on results section for
  Playwright assertions in plan 43-09
- Underscore _view-shell prefix prevents Next from treating it as a
  route segment
